Move MeshNormals tests to a separate file
